% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compo_new.R
\docType{data}
\name{compo_new}
\alias{compo_new}
\title{Data base of cultures between 2015 and 2018}
\format{A dataframe with 198 columns and 24 rows
\describe{
\item{X}{Name and year of the culture type, id}
\item{Year}{Year of study for the culture type}
\item{Culture}{Type of culture}
\item{Abia_sericea}{This column and the following present for each ok the 195 species
its relative abundance in each culture and each year}
}}
\source{
CEBC (Centre d'Etude Biologique de Chize)
}
\usage{
compo_new
}
\description{
This data base contains the relative abundance of pollinisators
in different types of culture fields.
}
\keyword{datasets}
