% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_an_eff.R
\name{moyenne}
\alias{moyenne}
\title{Effet indirect pour des interactions inter annuelles}
\usage{
moyenne(x, na_rm = FALSE)
}
\arguments{
\item{x}{Une matrice}
}
\value{
La fonction renvoie la valeur de l'indice de Muller pour chaque couple d'especes
}
\description{
Cette fonction calcule d'interactions entre especes d'un meme niveau trophique via un une espece d'un
niveau trophique different.Les especes en interactions sont independantes. Exemple:calcule l'indice de Muller pour
des interactions inter annuelles entre cultures via le partage de pollinisateurs.
}
\examples{
moyenne(c(2, 4))

}
