% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muller_inter.R
\name{muller_inter}
\alias{muller_inter}
\title{Effet indirect pour des interactions inter annuelles
Cette fonction calcule d'interactions entre especes d'un meme niveau trophique via un une espece d'un
niveau trophique different.Les especes en interactions sont independantes. Exemple:calcule l'indice de Muller pour
des interactions inter annuelles entre cultures via le partage de pollinisateurs.}
\usage{
muller_inter(
  ab_tab,
  start_year = 2015,
  start_crop = "Cereale",
  end_crop = "Colza",
  pol = "Abia_sericea"
)
}
\arguments{
\item{ab_tab}{a dataframe}

\item{start_year}{numeric}

\item{start_crop}{a character}

\item{end_crop}{a character}

\item{pol}{a character}
}
\value{
La fonction renvoie la valeur de l'indice de Muller pour une especes de pollinisateur pour un couple de
culture et une annee.
}
\description{
Effet indirect pour des interactions inter annuelles
Cette fonction calcule d'interactions entre especes d'un meme niveau trophique via un une espece d'un
niveau trophique different.Les especes en interactions sont independantes. Exemple:calcule l'indice de Muller pour
des interactions inter annuelles entre cultures via le partage de pollinisateurs.
}
\examples{
data(X)

muller_inter(
  ab_tab = X, start_year = 2015, start_crop = "Cereale",
  end_crop = "Colza", pol = "Abia_sericea"
)

}
