% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mullerinteran-package.R
\docType{package}
\name{mullerinteran-package}
\alias{mullerinteran}
\alias{mullerinteran-package}
\title{mullerinteran: Modify function PAC from package bipartite to calculate inter annual indirect efffects}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
