% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_an_eff.R
\name{moyenne}
\alias{moyenne}
\title{Moyenne d'un vecteur numerique}
\usage{
moyenne(x, na_rm = FALSE)
}
\arguments{
\item{x}{Un vecteur numerique sur lequel sera calculer la moyenne}

\item{na_rm}{Si TRUE supprime les NA (Defaut : FALSE)}
}
\value{
La fonction renvoie la moyenne du vecteur x.
}
\description{
Cette fonction calcule la moyenne d'un vecteur numerique en
omettant les valeurs manquantes si l'utilisateur le souhaite.
}
\examples{
moyenne(c(2, 4))

}
