% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mullerinteran-package.R
\docType{package}
\name{mullerinteran-package}
\alias{mullerinteran}
\alias{mullerinteran-package}
\title{mullerinteran: Modify function PAC from package bipartite to calculate inter annual indirect efffects}
\description{
The function calculate the interannual indirect effect between two crops for one pollinator species
}
\keyword{internal}
