% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.R
\docType{data}
\name{X}
\alias{X}
\title{Random dataset to run muller function}
\format{A dataframe with 4 lines and 6 columns
\describe{
\item{Year}{years}
\item{Culture}{random words figuring crop names}
\item{sit}{this column and the others figure the realtive abundance of virtual species}
}}
\source{
Random Number Generator (RNG): 1053
}
\usage{
X
}
\description{
Random dataset to run muller function
}
\keyword{datasets}
